requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langgraph>=0.2.14
langchain>=0.3.0
langchain-google-genai>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tavily-python>=0.3.0
structlog>=23.0.0
python-dotenv>=1.0.0
